add cli.py, core of the user interface.
cli.py contains the Commands class, which is the core of the applications
user interface. Commands acts as a wrapper around both argparse and
prompt_toolkit.
The root of the application commands are provided by cli.py as a
Command object called 'cli'. The 'quit' and 'source' commands are added
to the root command by cli.py