Warning + req force to add cname in delegated zone
This commit adds a warning message and requires the -force parameter
for the "host cname_add" command if the user tries to add a cname
in a delegated zone (i.e. one that Mreg doesn't control).
This is similar to the warning you get for "host add".

Background: https://rt.uio.no/Ticket/Display.html?id=4338328