Linting and formatting, semi-lax. (#172)
* Linting and formatting, semi-lax.

  - `tox` now runs `black` and `ruff` on the project (the environment is called `lint`).
  - tox runs `ruff` with --fix to resolve fixable issues.
  - There is no pre-commit hook or CI integration as of yet.

### Code changes

To adhere to even semi-lax linting, there were changes required in the code. These are:

- Migrated use of python builtin identifiers (`type`,  `id`, and `print`) to more appropriate names.
- Some cases of long string literals were using implicit concatenation, which is now explicit (by adding a set of parentheses).

### Outstanding changes with current settings

- `print` in history.py (shadowing builtin) was not fixed as it needs a bit more debugging to figure out exactly how it is used.

### Notes

There is still a lot to do with regards to linting, formatting, and typing. This PR does not seek to resolve these issues, but attempts to create a baseline of requirements for the codebase to adhere to.