Major refactoring.
  - Split util.py into modules under mreg_cli.utilities.
  - Move command modules in under mreg_cli.comands.
  - Use decorators for commands instead of following callbacks with command additions.
  - Create a proper configuration module (config.py).
  - Set up a help command.
  - Update README.md to match reality.
  - Remove the broken history implemementation.

Command modules create a command class that inherits from mreg_cli.commands.base.BaseCommand. This class is initialized with a mreg_cli.commands.registry.CommandRegistry that returns the decorator used. See the current commands for examples.

Supports input from a configuration file, environment, and comand line and prioritises the keys in that order.

A help command is now available. At the root, it does the same as `-h`, but it also allows for subcommands. Currently supported subcommands:

  - `filtering` provides help about the filtering operators, with examples.
  - `configuration` shows the configuration options currently being employed.