Verified behaviour for all group commands
I have looked at the new output and api calls for the "group" commands
in the testsuite and it looks good to me.
Fixed a bug in history.py concerning how history is printed when owners
or hosts are added/removed from hostgroups.
Made a minor change to the datetime regex in diff.py to handle the extra
space that comes before a single digit date.