Add a check for frozen/force to "host add"
We forgot to check if the network was frozen in "host add". Added a
check for that in this commit, but we'll probably have to do
something similar in a_aaaa.py.

Went through more command output and verified it. Mostly network
commands this time.

Modified diff.py to not care about length of white space. Since ip
addresses can vary between runs, formatted output can vary in length,
spesifically how much whitespace is added.